requirements.txt
cffi==1.16.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
kiwisolver==1.4.5
packaging==23.2
pillow==10.2.0
PyAudio==0.2.14
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pyusb==1.2.1
six==1.16.0
tk==0.1.0
